requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.26.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
openai==0.27.0
python-dotenv==0.19.0
gunicorn==20.1.0
python-dateutil==2.8.2 